Rework page.java
Firstly, multi-lined variable assignments is okay, but it's a bit dirty looking

Some variables weren't needed such as pageAccessToken and assigning JSON_URL.
These variables are only used once in the constructor and never again.

Rename a few variables, c should be facebookClient and not single letters.

Rework while function, it can be replaced with some logic and not variables.