Rework project
This is a giant commit so bare with me... The structure of the code seemed a
little off, the logic of it all was solid but error catching wasn't solid
enough. There shouldn't be so many "throws error" declarations. This rework
mostly attempted to conver the object classes to static util classes while
also having a nice trail of "null" values to determine if the code failed.

In the util classes, the functions in them will return null when an issue
happens for some reason or another. This will travel back up to the main
class where a timeout now occurs. JSON errors are handled in JSONParser
and Facebook errors are handled in FBHandler. Workshop is a new class
to easily package some information to store into a list. Since the
strings in the class Workshop are final, the variables can be public.

The main class now pieces together everything and posts. It contains
some logic for a timeout functionality, but errors should be handled
a bit more... smart in the future, not sure how yet.